requirements.txt
cached-property==1.3.0
click==6.7
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
docker-compose==1.9.0
docker-py==1.10.6
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
Flask==0.12
furl==0.5.6
itsdangerous==0.24
Jinja2==2.9.4
jsonschema==2.5.1
Markdown==2.6.7
MarkupSafe==0.23
orderedmultidict==0.7.11
path.py==10.0
pika==0.10.0
py==1.4.32
pytest==3.0.5
PyYAML==3.12
requests==2.11.1
six==1.10.0
texttable==0.8.7
websocket-client==0.40.0
Werkzeug==0.11.15
